move demo sub components to own dir
