+++ sales-coach-avatar/src/app/demo/_views/demo/_components/DemoDescription.tsx
