+++ sales-coach-avatar/src/app/demo/_views/demo/_components/index.ts
@@ -1,0 +1,2 @@
+export * from './AvatarContainer';
+export * from './DemoDescription';
